refactor(module:auto-complete): remove NZ_AUTOCOMPLETE_VALUE_ACCESSOR
