+++ components/auto-complete/error.ts
@@ -1,0 +1,15 @@
+/**
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
+ */
+
+/**
+ * @note Internally used, please do not use it
+ */
+export function getNzAutocompleteMissingPanelError(): Error {
+  return Error(
+    'Attempting to open an undefined instance of `nz-autocomplete`. ' +
+      'Make sure that the id passed to the `nzAutocomplete` is correct and that ' +
+      "you're attempting to open it after the ngAfterContentInit hook."
+  );
+}
